Claude/add ingredient dropdown 48bm h

+++ DelikatessenDrehbuch/Models/Quantity.cs
@@ -3,8 +3,8 @@
     public class Quantity
     {
         public int Id { get; set; }
-        public float Quantitys { get; set; }
-       
+        public double Quantitys { get; set; }
+
 
     }
 }
